+++ examples/hello_world.el
@@ -1,0 +1,3 @@
+// 我的第一个Evil Lang程序
+// My first Evil Lang program
+print("Hello, Evil World!");
