+++ src/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+# Evil Lang - 一个简单的编程语言解释器 / A Simple Programming Language Interpreter
+# Author: Evil0ctal
+# Date: 2025-05-04
+
+# 空的__init__.py文件，使src目录成为一个Python包
+# Empty __init__.py file, making the src directory a Python package
